createQueryNamespace {ResultModelManager}	R Documentation

Create query namespace

Description

Create a QueryNamespace instance from either a connection handler or a connectionDetails object Allows construction with various options not handled by QueryNamespace$new

Note - currently not supported is having multiple table prefixes for multiple table namespaces

Usage

createQueryNamespace(
  connectionDetails = NULL,
  connectionHandler = NULL,
  usePooledConnection = FALSE,
  tableSpecification = NULL,
  resultModelSpecificationPath = NULL,
  tablePrefix = "",
  snakeCaseToCamelCase = TRUE,
  ...
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

connectionHandler
ResultModelManager ConnectionHandler or PooledConnectionHandler instance

usePooledConnection
Use Pooled database connection instead of standard DatabaseConnector single connection.

tableSpecification
Table specfication data.frame

resultModelSpecificationPath
(optional) csv file or files for tableSpecifications - must conform to table spec format.

tablePrefix
String to prefix table names with - default is empty string

snakeCaseToCamelCase
convert snakecase results to camelCase field names (TRUE by default)

...
Elipsis - use for any additional string keys to replace

[Package ResultModelManager version 0.4.0 Index]