deleteAllRowsForPrimaryKey {ResultModelManager}	R Documentation

Delete results rows for primary key values from database server tables

Description

Delete results rows for primary key values from database server tables

Usage

deleteAllRowsForPrimaryKey(connection, schema, tableName, keyValues)

Arguments

connection
DatabaseConnector connection instance

schema
The schema on the postgres server where the results table exists

tableName
Database table name

keyValues
Key values of results rows to be deleted

Details

Only PostgreSQL servers are supported.

[Package ResultModelManager version 0.3.0 Index]