unzipResults {ResultModelManager}	R Documentation

Unzips a results.zip file and enforces standards required by uploadResults

Description

This function will unzip the zipFile to the resultsFolder and assert that the file resultsDataModelSpecification.csv exists in the resultsFolder to ensure that it will work with uploadResults

Usage

unzipResults(zipFile, resultsFolder)

Arguments

zipFile
The location of the .zip file that holds the results to upload

resultsFolder
The folder to use when unzipping the .zip file. If this folder does not exist, this function will attempt to create the folder.

[Package ResultModelManager version 0.3.0 Index]