PooledConnectionHandler {ResultModelManager}	R Documentation

Pooled Connection Handler

Description

Transparently works the same way as a standard connection handler but stores pooled connections. Useful for long running applications that serve multiple concurrent requests.

Super class

ResultModelManager::ConnectionHandler -> PooledConnectionHandler

Methods

Public methods

PooledConnectionHandler$new()

PooledConnectionHandler$initConnection()

PooledConnectionHandler$dbms()

PooledConnectionHandler$closeConnection()

PooledConnectionHandler$queryFunction()

PooledConnectionHandler$executeFunction()

PooledConnectionHandler$clone()

Inherited methods

ResultModelManager::ConnectionHandler$dbIsValid()

ResultModelManager::ConnectionHandler$executeSql()

ResultModelManager::ConnectionHandler$finalize()

ResultModelManager::ConnectionHandler$getConnection()

ResultModelManager::ConnectionHandler$queryDb()

ResultModelManager::ConnectionHandler$renderTranslateSql()

ResultModelManager::ConnectionHandler$tbl()

Method new()

Usage

PooledConnectionHandler$new(...)

Arguments

...

Elisis @seealsoConnectionHandler initialize pooled db connection

Method initConnection()

Overrides ConnectionHandler Call get dbms

Usage

PooledConnectionHandler$initConnection()

Method dbms()

Get the dbms type of the connection Close Connection

Usage

PooledConnectionHandler$dbms()

Method closeConnection()

Overrides ConnectionHandler Call query Function

Usage

PooledConnectionHandler$closeConnection()

Method queryFunction()

Overrides ConnectionHandler Call. Does not translate or render sql.

Usage

PooledConnectionHandler$queryFunction(
  sql,
  snakeCaseToCamelCase = self$snakeCaseToCamelCase
)

Arguments

sql

sql query string

snakeCaseToCamelCase

(Optional) Boolean. return the results columns in camel case (default) query Function

Method executeFunction()

Overrides ConnectionHandler Call. Does not translate or render sql.

Usage

PooledConnectionHandler$executeFunction(sql)

Arguments

sql

sql query string

Method clone()

The objects of this class are cloneable with this method.

Usage

PooledConnectionHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package ResultModelManager version 0.3.0 Index]